Box nested fan callbacks without redundant ownership copies
