+++ crates/almide-spine/tests/golden/spec-run-manifest.txt
@@ -143,6 +143,7 @@ e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855	1	spec/wasm_cro
 e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855	1	spec/wasm_cross/fan_map_err.almd
 e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855	1	spec/wasm_cross/fan_map_inline_err.almd
 f402052ae966bbed45c70cede8be710a59c85b5f7dda6b90e62cc368e323022b	0	spec/wasm_cross/fan_map_inline_lambda.almd
+f686262d57bdafccc8a64b84668c7439c66e2e230bc2c3b20098efbcf1e348dd	0	spec/wasm_cross/fan_map_nested_combinators.almd
 fc2c7cdf95fa30121ceb5a688570f30af7534bf5d9bb1a497393537f444d5dd5	0	spec/wasm_cross/fan_map_parallel_scalar.almd
 c71631ce4f3e018b5dab2b27afa91c695e468fd33be6b5146a66c84c1c79f5e9	0	spec/wasm_cross/fan_prefetch_fs.almd
 c64f10d8ccfe32e86580dc864d9466053132ef751edc2c164c3d44f9ef5b8b97	0	spec/wasm_cross/fan_pure_thunks.almd
