+++ crates/almide-syntax/tests/golden/spec-ast-manifest.txt
@@ -746,6 +746,7 @@ b3665b340f097abf85dd0d1b87289d685a21cd87d1e140cdf1bac91972ee7d76	spec/wasm_cross
 fe024f07bd1fee7d3eac86aa26d09de4ab35f78451f6c1b945a04dc63b36f3a4	spec/wasm_cross/fan_map_err.almd
 7736e18ed6128b64cfe600fbd54737534ef6da6b49e816917cae1ee0f3ebec40	spec/wasm_cross/fan_map_inline_err.almd
 a32a3e07d581ade3f2fd7dc93f664470db34423ee561d6ab8d3cbc7181525113	spec/wasm_cross/fan_map_inline_lambda.almd
+4804aa1f6b9c8bc555baff286aeb76f83d1cb4502d648385b1dae39eae0baad0	spec/wasm_cross/fan_map_nested_combinators.almd
 573b8c014fe62e704b8b0fca1c3a82d588b87b171f72172b4f6e669a5f623eb2	spec/wasm_cross/fan_map_parallel_scalar.almd
 c2beaa668f3366b057d02effbab6542411dba07a551971db0d7c409464c6350f	spec/wasm_cross/fan_prefetch_fs.almd
 4866c7dc43750388ee32f713a8560a3e3a29967c435b3e280c9b643de8a91b61	spec/wasm_cross/fan_pure_thunks.almd
