+++ docs/ARCHITECTURE.md
@@ -253,7 +253,7 @@ The Wgsl arm is the four rows marked W. Class:
 | 23 | `DecodeSlotHint` | `pass_decode_slot_hint.rs` | Rust | Rust by design | a derived `T.decode`'s borrowed field lookups carry their declaration index (`@codec_slots`), rendered as `almide_rt_value_field_ref_at` (#1679); the runtime tries `pairs[i]` before the scan, errors unchanged (C-084) | n/a (the wasm leg reads neither the attribute nor the `_at` symbol) |
 | 23b | `DecodeErrFrame` | `pass_decode_err_frame.rs` | Rust | Rust by design | a derived `T.decode`'s per-field error frame (`T___erratw_*`, #1675) becomes `.map_err(..)` on the field's own result, so the success path is a plain `?` with no payload clone and no key allocation (#2050) | n/a (the wasm leg lowers the worker itself) |
 | 24 | `Peephole` | `pass_peephole.rs` | all | optimizer | idiomatic list loops → `ListSwap`/`ListReverse`/`ListRotateLeft`/`ListCopySlice` nodes | **no lowering for those nodes** — see E |
-| 25 | `RustLowering` | `pass_rust_lowering.rs`, `pass_rust_lowering_fan.rs`, `pass_fan_local_state.rs` | Rust | Rust by design | push optimization, borrow index lift; routes `fan.map` and list ops under `fan { }` with a pure lambda over Send-safe scalars to the thread-per-core runtime twins (`almide_rt_fan_map_par`, `almide_rt_list_par_*`; `ALMIDE_FAN_SEQUENTIAL=1` ablates, #2044) | n/a |
+| 25 | `RustLowering` | `pass_rust_lowering.rs`, `pass_rust_lowering_stmts.rs`, `pass_rust_lowering_fan.rs`, `pass_fan_local_state.rs` | Rust | Rust by design | push optimization, borrow index lift; routes `fan.map` and list ops under `fan { }` with a pure lambda over Send-safe scalars to the thread-per-core runtime twins (`almide_rt_fan_map_par`, `almide_rt_list_par_*`; `ALMIDE_FAN_SEQUENTIAL=1` ablates, #2044) | n/a |
 | 26 | `FanLowering` | `pass_fan_lowering.rs` (wrapper in `pass.rs`) | all, W | enabler | strip auto-try from fan spawn closures | own fan lowering (`fan.rs`) |
 | 27 | `NormalizeRuntimeCalls` | `pass_normalize_runtime_calls.rs` | Rust | Rust by design | legacy `Named { almide_rt_* }` → `RuntimeCall` | n/a |
 | 28 | `IrLinkFlatten` | `pass_ir_link_flatten.rs` | Rust | Rust by design | flatten modules into the root for the walker | keeps modules, qualified names |
